remove bnoordhuis from lists (#67)
Ben has resigned from TSC, so removing from the TSC and CTC lists. I
made the judgment call that he should be removed from the security list
as well as everyone else on that list is TSC and they can always loop
him in if need be. Not feeling too strongly about that though if there's
strong opposition.